Service Worker
Introduction to the core concepts
of the service worker, some life-cycles
methods like register, installing and activating.

Demonstration of the fetch api with interception
of the requests of our app and how we can rewrite
the responses of the request to something else
other the original response.

listen to the beforeinstallpromt event to check
if the user wants to install the app.